+++ src/ONTraC/analysis/data.py
@@ -213,8 +213,11 @@ class AnaData:
         params = read_yaml_file(f'{options.preprocessing_dir}/samples.yaml')
         self.rel_params = get_rel_params(options, params)
         # save the original Cell ID
-        self.cell_id = pd.read_csv(f'{options.preprocessing_dir}/meta_data.csv', usecols=['Cell_ID']).set_index('Cell_ID')
-        self.meta_data = pd.read_csv(f'{options.preprocessing_dir}/meta_data.csv')
+        self.cell_id = pd.read_csv(f'{options.preprocessing_dir}/meta_data.csv',
+                                   usecols=['Cell_ID']).set_index('Cell_ID', inplace=False)
+        self.meta_data = pd.read_csv(f'{options.preprocessing_dir}/meta_data.csv').set_index('Cell_ID', inplace=False)
+        if 'Cell_Type' in self.meta_data.columns:
+            self.cell_id['Cell_Type'] = self.meta_data['Cell_Type'] = self.meta_data['Cell_Type'].astype('category')
 
     @property
     def train_loss(self):
@@ -245,7 +248,7 @@ class AnaData:
                 continue
             feature_file = f"{sample['Features'][:-27]}_Raw_CellTypeComposition.csv.gz"
             if not os.path.isfile(feature_file):
-                    continue
+                continue
             cell_type_composition_df = pd.read_csv(feature_file, header=None)
             cell_type_composition_df.columns = self.cell_type_codes.loc[np.arange(cell_type_composition_df.shape[1]),
                                                                         'Cell_Type'].tolist()  # type: ignore
@@ -259,7 +262,8 @@ class AnaData:
         self._NT_score = data_df[['sample', 'x', 'y', 'Niche_NTScore', 'Cell_NTScore']]
         if not data_2_df.empty:
             data_2_df = self.cell_id.join(data_2_df)
-            self._adjust_cell_type_composition = data_2_df[self.cell_type_codes['Cell_Type'].tolist() + ['sample', 'x', 'y']]
+            self._adjust_cell_type_composition = data_2_df[self.cell_type_codes['Cell_Type'].tolist() +
+                                                           ['sample', 'x', 'y']]
 
     @property
     def cell_type_composition(self) -> DataFrame: