+++ src/ONTraC/optparser/_preprocessing.py
@@ -14,6 +14,25 @@ IO_OPTIONS = ['input', 'preprocessing_dir']
 # ------------------------------------
 # Functions
 # ------------------------------------
+def add_preprocessing_options_group(optparser: OptionParser) -> None:
+    """
+    Add preprocessing options group to optparser.
+    :param optparser: OptionParser object.
+    :return: OptionGroup object.
+    """
+    # preprocessing options group
+    group_preprocessing = OptionGroup(optparser, "Preprocessing")
+    group_preprocessing.add_option(
+        '--resolution',
+        dest='resolution',
+        type=float,
+        default=10.0,
+        help=
+        'Resolution for leiden clustering. Used for clustering cells into cell types when gene expression data is provided. Default is 10.0.'
+    )
+    optparser.add_option_group(group_preprocessing)
+
+
 def add_niche_net_constr_options_group(optparser: OptionParser) -> None:
     """
     Add niche network construction options group to optparser.
@@ -90,6 +109,17 @@ def validate_niche_net_constr_options(optparser: OptionParser, options: Values) 
             sys.exit(1)
 
 
+def write_preprocessing_memo(options: Values):
+    """Write preprocessing memos to stdout.
+
+    Args:
+        options: Options object.
+    """
+
+    # print parameters to stdout
+    info('      -------- preprocessing options -------      ')
+    info(f'resolution: {options.resolution}')
+
 def write_niche_net_constr_memo(options: Values):
     """Write niche network construction memos to stdout.
 
@@ -121,6 +151,9 @@ def prepare_preprocessing_optparser() -> OptionParser:
     # I/O options group
     add_IO_options_group(optparser=optparser, io_options=IO_OPTIONS)
 
+    # preprocessing options group
+    add_preprocessing_options_group(optparser)
+
     # niche network construction options group
     add_niche_net_constr_options_group(optparser)
 
@@ -141,6 +174,7 @@ def opt_preprocessing_validate(optparser) -> Values:
     # print parameters to stdout
     info('------------------ RUN params memo ------------------ ')
     write_io_options_memo(options, IO_OPTIONS)
+    write_preprocessing_memo(options)
     write_niche_net_constr_memo(options)
     info('--------------- RUN params memo end ----------------- ')